requirement.txt
anyio==3.6.2
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
python-telegram-bot==20.2
requests==2.29.0
rfc3986==1.5.0
schedule==1.2.0
sniffio==1.3.0
soupsieve==2.4.1
urllib3==1.26.15
